Moved render release card method to module
